backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
python-dotenv
pymysql